moves logic for getting repos and project columns into base runner
